remove LogExporter
